requirements.txt
async-generator==1.10
attrs==19.3.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.14.0
click==8.0.3
constantly==15.1.0
cryptography==3.3.2
cssselect==1.1.0
Flask==2.0.2
Flask-Pydantic==0.9.0
google-cloud-bigquery==3.4.1
google-cloud-tasks==2.11.0
gunicorn==20.1.0
h11==0.12.0
hyperlink==19.0.0
idna==2.9
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.5
outcome==1.1.0
pandas==1.3.5
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
queuelib==1.5.0
Scrapy==2.0.1
selenium==4.1.0
service-identity==18.1.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.7.0
typing-extensions==4.0.1
urllib3==1.26.5
w3lib==1.21.0
Werkzeug==2.0.2
wsproto==1.0.0
zope.interface==5.0.0